% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qiitaddin-package.r
\docType{package}
\name{qiitaddin}
\alias{qiitaddin}
\alias{qiitaddin-package}
\title{An RStudio Addin To Post To Qiita}
\description{
Knit an R Markdown file and post the result Markdown file to Qiita
}

