% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qiitaddin.r
\name{qiitaddin_knit}
\alias{qiitaddin_knit}
\title{Knit And Post A Rmarkdown Document to Qiita}
\usage{
qiitaddin_knit(input = NULL)
}
\arguments{
\item{input}{path to a .Rmd file. If \code{NULL}, use the activedocument.}
}
\description{
Knit And Post A Rmarkdown Document to Qiita
}
