% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gyazo.r
\name{upload_image_gyazo}
\alias{upload_image_gyazo}
\title{Upload an image to Gyazo}
\usage{
upload_image_gyazo(file)
}
\arguments{
\item{file}{Path to an image file.}
}
\value{
A URL of the uploaded image.
}
\description{
Uploads a given image to Gyazo.
}
