% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imgur.r
\name{upload_image_imgur}
\alias{upload_image_imgur}
\title{Upload an image to Imgur}
\usage{
upload_image_imgur(file)
}
\arguments{
\item{file}{Path to an image file.}
}
\value{
A URL of the uploaded image.
}
\description{
Uploads a given image file to Imgur by imguR package.
}
