% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qiitaddin.r
\name{qiitaddin_knit}
\alias{qiitaddin_knit}
\title{Publish R Markdown Document to Qiita}
\usage{
qiitaddin_knit(input = NULL)
}
\arguments{
\item{input}{path to a .Rmd file. If \code{NULL}, use the activedocument.}
}
\description{
Knit and post a given R Markdown file to Qiita. Images will be uploaded to Imgur or Gyazo.
}
